Appify

Appify

appify(f, inputs = list(), id_postfix = NULL, input_check = TRUE,
  plot_height = 600, position_flip = FALSE)

Arguments

f
A function from your package or a function written in your Rmd document.

inputs
Inputs to your function. A form will be created for them.

id_postfix
Advanced users may wish to set the html-id postfix. The default value is NULL. appify will then create a random postfix.

input_check
Should inputs be checked against function inputs (default: TRUE)?

plot_height
Pixel value for the plot height. The default value is 600. So, the app including the form will comfortably fit on a laptop screen.

position_flip
By default (FALSE) the plot will be placed below the input form.

Value

html and javascript code

Examples

# Appify returns html. When included in a Rmarkdown document this will be
# display as a form and a canvas.
rnorm_plot <- function(n, mean, sd) {
  plot(rnorm(n = n, mean = mean, sd = sd))
}

appify(
  f = rnorm_plot,
  inputs = list(
    n = number("Sample Size"),
    mean = number("Distribution Mean"),
    sd = number("Distribution Standard Deviation")
  )
)

#> <div>
#>   <div>
#>     <form id="form-eRFiayzpMOTjbqukuduYpJMkYMcBKJbl" class="well">
#>       <fieldset class="row">
#>         <div class="col-sm-4">
#>           <label>Sample Size</label>
#>           <input id="input-n-eRFiayzpMOTjbqukuduYpJMkYMcBKJbl" class="form-control" type="number" min="0" max="100"/>
#>         </div>
#>         <div class="col-sm-4">
#>           <label>Distribution Mean</label>
#>           <input id="input-mean-eRFiayzpMOTjbqukuduYpJMkYMcBKJbl" class="form-control" type="number" min="0" max="100"/>
#>         </div>
#>         <div class="col-sm-4">
#>           <label>Distribution Standard Deviation</label>
#>           <input id="input-sd-eRFiayzpMOTjbqukuduYpJMkYMcBKJbl" class="form-control" type="number" min="0" max="100"/>
#>         </div>
#>       </fieldset>
#>       <button id="button-eRFiayzpMOTjbqukuduYpJMkYMcBKJbl" class="btn btn-primary" type="button">Submit</button>
#>     </form>
#>     <div style="height: 600px;" class="well" id="canvas-eRFiayzpMOTjbqukuduYpJMkYMcBKJbl"></div>
#>   </div>
#>   <script type="text/javascript" charset="utf-8">$("#button-eRFiayzpMOTjbqukuduYpJMkYMcBKJbl").click(function(){
#> var req = $("#canvas-eRFiayzpMOTjbqukuduYpJMkYMcBKJbl").rplot("rnorm_plot", {
#>     n : (function(){x = $("#input-n-eRFiayzpMOTjbqukuduYpJMkYMcBKJbl").val(); return(isNaN(x) ? x : parseFloat(x))})(),
#> mean : (function(){x = $("#input-mean-eRFiayzpMOTjbqukuduYpJMkYMcBKJbl").val(); return(isNaN(x) ? x : parseFloat(x))})(),
#> sd : (function(){x = $("#input-sd-eRFiayzpMOTjbqukuduYpJMkYMcBKJbl").val(); return(isNaN(x) ? x : parseFloat(x))})()
#> });
#> 
#> //Error
#> req.fail(function(){
#>     alert("Error in R function call: " + req.responseText);
#> });
#> });</script>
#> </div>