appifyr: Create Web Applications from R Functions

The appifyr package enables R users to create interactive web applications using OpenCPU as a backend. It provides a bridge between R's analytical capabilities and web interfaces by automatically generating HTML forms for function parameters, connecting these inputs to OpenCPU API calls, and displaying the function outputs.

To learn more about appifyr, start with the vignettes: `browseVignettes(package = "appifyr")`

Key Functions

appify: Create a web app from an R function

inp_text: Create a text input element

inp_number: Create a numeric input element

inp_dropdown: Create a dropdown/select input element

build_app: Build and install the application

See also

Author

Maintainer: Reto Wyss reto.wyss.ch@gmail.com [copyright holder]

Other contributors:

Michael Rustler michael.rustler@kompetenz-wasser.de (ORCID) [contributor]